requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
paho-mqtt==2.1.0
python-multipart==0.0.17
websockets==13.1
jinja2==3.1.4
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
pytest~=8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.2
aiohttp==3.10.10
mock==5.1.0
